+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Use devRoot to resolve MIG device nodes.
 * Fix bug in determining default nvidia-container-runtime.user config value on SUSE-based systems.
 * Add `crun` to the list of configured low-level runtimes.
+* Added support for `--ldconfig-path` to `nvidia-ctk cdi generate` command.
 
 * [toolkit-container] Bump CUDA base image version to 12.3.1.